# address some linting issues, and small refactor, performance optimizations